Split getInitDataFromPro y two function in ContentProtection and PSSH
