+++ lib/dash/content_protection.js
@@ -322,6 +322,31 @@ shaka.dash.ContentProtection = class {
   }
 
   /**
+   * Gets a PlayReady initData from a content protection element
+   * containing a PlayReady Pro Object
+   *
+   * @param {shaka.dash.ContentProtection.Element} element
+   * @return {?Array.<shaka.extern.InitDataOverride>}
+   */
+  static getInitDataFromPro_(element) {
+    const proNode = shaka.util.XmlUtils.findChildNS(
+        element.node, 'urn:microsoft:playready', 'pro');
+    if (!proNode) {
+      return null;
+    }
+    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
+    const data = Uint8ArrayUtils.fromBase64(proNode.textContent);
+    const pssh = shaka.util.Pssh.getPSSHFromMsprPro(data);
+    return [
+      {
+        initData: pssh,
+        initDataType: 'cenc',
+        keyId: element.keyId,
+      },
+    ];
+  }
+
+  /**
    * Creates DrmInfo objects from the given element.
    *
    * @param {Array.<shaka.extern.InitDataOverride>} defaultInit
@@ -346,7 +371,7 @@ shaka.dash.ContentProtection = class {
             !element.init || element.init.length,
             'Init data must be null or non-empty.');
 
-        const proInitData = shaka.util.Pssh.getInitDataFromPro(element);
+        const proInitData = ContentProtection.getInitDataFromPro_(element);
         const initData = element.init || defaultInit || proInitData;
         const info = ManifestParserUtils.createDrmInfo(keySystem, initData);
         const licenseParser = licenseUrlParsers.get(keySystem);
